requirements.txt
Pillow
numpy
transformers
diffusers
omegaconf
